+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 - Update context for Illinois Assistant [#3410](https://github.com/rokwire/illinois-app/issues/3410)
+### Fixed
+- Preserve user favorites stored in Core account data [#3412](https://github.com/rokwire/calendar-building-block/issues/3412).
 
 ## [5.0.27] - 2023-07-13
 ### Changed